+++ src/sql/magic_plot.py
@@ -14,6 +14,7 @@ except ImportError:
 
 from sql import plot
 from sql.command import SQLPlotCommand
+from sql import util
 
 
 @magics_class
@@ -68,6 +69,8 @@ class SqlPlotMagic(Magics, Configurable):
             )
 
         if cmd.args.line[0] in {"box", "boxplot"}:
+            util.is_table_exists(cmd.args.table, with_=cmd.args.with_)
+
             return plot.boxplot(
                 table=cmd.args.table,
                 column=column,
@@ -76,6 +79,8 @@ class SqlPlotMagic(Magics, Configurable):
                 conn=None,
             )
         elif cmd.args.line[0] in {"hist", "histogram"}:
+            util.is_table_exists(cmd.args.table, with_=cmd.args.with_)
+
             return plot.histogram(
                 table=cmd.args.table,
                 column=column,